requirements.txt
matplotlib
ipykernel
notebook
seaborn
pyyaml
pandas
numpy
scipy
